+++ include/sensor_lorawan.h
@@ -25,7 +25,7 @@ typedef struct {
     /* Whether the LoRaWAN is enabled. */
     uint8_t is_lorawan_enabled;
     /* LoRaWAN frequency. */
-    uint32_t lorawan_frequency;
+    uint8_t lorawan_frequency;
     /* LoRaWAN Device. */
     const struct device *lora_dev;
     /* LoRaWAN uplink class. */
